WIP refactoring - simplifications
